requirements.txt
python-dotenv
channels
amqp==2.5.1
appnope==0.1.0
appointmentguru-helpers==0.3
astroid==2.2.2
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
backcall==0.1.0
beautifulsoup4==4.8.0
billiard==3.6.1.0
boto==2.48.0
boto3==1.9.198
botocore==1.12.198
celery==4.4.0rc3
certifi==2019.3.9
chardet==3.0.4
Click==7.0
codacy-coverage==1.3.11
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
cssutils==1.0.2
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1.11
django-admin==1.3.2
django-allauth==0.39.1
django-autoslug-iplweb==1.9.4
django-bootstrap4==0.0.7
django-cc==0.2.3
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-excel-response2==2.0.8
django-filter==1.0.4
django-inlinecss==0.2.0
django-payfast==0.3
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-six==1.0.4
django-social-share==1.3.2
django-storages==1.6.3
django-templated-mail==1.1.1
django-templatetags==1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
djet==0.2.2
docutils==0.14
Faker==2.0.0
Flask==1.1.1
future==0.17.1
#GDAL
gunicorn==19.7.1
html5lib==1.0.1
hypothesis==4.32.3
idna==2.8
importlib-metadata==0.19
ipdb==0.12.1
ipython==7.7.0
ipython-genutils==0.2.0
isort==4.3.10
itsdangerous==1.1.0
itypes==1.1.0
jedi==0.14.1
Jinja2==2.10.1
jmespath==0.9.4
kombu==4.6.4
kong-oauth==0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
mock==3.0.5
more-itertools==7.2.0
mysql-connector-python==8.0.18
numpy==1.18.1
oauthlib==3.0.1
packaging==19.1
parso==0.5.1
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pluggy==0.12.0
prompt-toolkit==2.0.9
protobuf==3.11.1
psycopg2-binary==2.8.3
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycoin==0.90.20190728
Pygments==2.4.0
PyJWT==1.7.1
pylint==2.3.1
pymessenger==0.0.7.0
PyMySQL==0.9.3
pynliner==0.8.0
pyparsing==2.4.2
pytest==5.0.1
pytest-cov==2.7.1
pytest-django==3.5.1
pytest-pythonpath==0.7.3
python-bitcoinrpc==1.0
python-dateutil==2.8.0
python-decouple==3.1
python-social-auth==0.2.12
python3-openid==3.1.0
pytz==2018.5
requests==2.21.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
responses==0.10.6
s3transfer==0.2.1
screen==1.0.1
selenium==3.141.0
sentry-sdk==0.6.1
six==1.12.0
slackclient==1.0.6
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.2
sqlparse==0.2.4
stripe==2.27.0
text-unidecode==1.2
traitlets==4.3.2
typed-ast==1.3.1
uritemplate==3.0.0
urllib3==1.24.2
vine==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.15.5
whitenoise==3.3.1
wrapt==1.11.1
xlwt==1.3.0
zipp==0.5.2